models renamed to model
- Removed model folder from client so have a flat structure
