+++ src/client/Entity.ts
