+++ src/provider/model/NotificationCenterAPI.ts
@@ -1,6 +1,6 @@
 import {ProviderIdentity} from 'openfin/_v2/api/interappbus/channel/channel';
 import {INotification} from './INotification';
-import {SenderInfo} from '../../client/model/Notification';
+import {SenderInfo} from '../../client/Notification';
 
 export interface NotificationCenterEventMap {
     'notificationCreated': INotification & SenderInfo;
