+++ src/provider/view/components/Notification/Notification.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import {NotificationTime} from '../NotificationTime/NotificationTime';
 import {NotificationCenterAPI} from '../../../model/NotificationCenterAPI';
 import {Button} from '../Button/Button';
-import {NotificationTypes} from '../../../../client/model/Notification';
+import {NotificationTypes} from '../../../../client/Notification';
 import {INotification} from '../../../model/INotification';
 declare var window: Window & {openfin: {notifications: NotificationCenterAPI}};
